refactor: mixin classes
