+++ packages/discordx/src/classes/index.ts
@@ -1,4 +1,4 @@
-export * from "./ApplicationCommandMixin.js";
-export * from "./ApplicationGuildMixin.js";
 export * from "./DefaultPermissionResolver.js";
+export * from "./Mixin/ApplicationCommandMixin.js";
+export * from "./Mixin/ApplicationGuildMixin.js";
 export * from "./SimpleCommandMessage.js";
